store, jsonutil: move store.getStructFields to jsonutil.StructFields
It seems something like StructFields is more generally useful, so this
moves it to jsontuil so it can be reused.
